+++ src/tests/src/libs/antares/study/constraint/test_constraint.cpp
@@ -13,7 +13,7 @@
 #include "antares/study/area/area.h"
 #include <antares/study/study.h>
 #include <filesystem>
-#include "utils.h"
+#include <files-system.h>
 
 using namespace Antares::Data;
 namespace fs = std::filesystem;
