+++ src/tests/src/libs/antares/study/constraint/test_group.cpp
@@ -9,7 +9,7 @@
 #include <fstream>
 #include <memory>
 #include "antares/study/study.h"
-#include "utils.h"
+#include <files-system.h>
 
 using namespace Antares::Data;
 namespace fs = std::filesystem;
