+++ src/tests/src/libs/antares/study/parts/hydro/test-hydro-series.cpp
@@ -6,7 +6,7 @@
 #include <study.h>
 #include <LoadingError.hpp>
 #include <matrix.h>
-#include <utils.h>
+#include <files-system.h>
 
 #define SEP "/"
