+++ src/tests/src/libs/antares/study/parts/hydro/test-hydroreader-class.cpp
@@ -5,7 +5,7 @@
 #include <boost/test/included/unit_test.hpp>
 #include <study.h>
 #include <matrix.h>
-#include <utils.h>
+#include <files-system.h>
 
 #define SEP "/"
