+++ src/tests/src/solver/simulation/test-store-timeseries-number.cpp
@@ -11,7 +11,7 @@
 #include <fstream>
 #include "timeseries-numbers.h"
 #include "BindingConstraintsTimeSeriesNumbersWriter.h"
-#include "utils.h"
+#include <files-system.h>
 #include <antares/writer/writer_factory.h>
 #include <antares/writer/result_format.h>
 #include <antares/benchmarking/DurationCollector.h>
