+++ src/tests/src/solver/simulation/test-time_series.cpp
@@ -9,7 +9,7 @@
 #include <antares/study/study.h>
 #include <filesystem>
 #include <fstream>
-#include "utils.h"
+#include <files-system.h>
 
 using namespace Antares::Solver;
 using namespace Antares::Data;
