+++ src/tests/src/utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(test_utils_unit
-        utils.cpp
-        utils.h
+        files-system.cpp
+        files-system.h
         )
 
 target_include_directories(
