+++ src/tests/src/utils/files-system.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "files-system.h"
 
 namespace fs = std::filesystem;
 
@@ -62,8 +62,8 @@ void removeFolder(stringT& path, stringT& folder_name)
         try
         {
             fs::remove_all(folder_path);
-            std::cout << "Folder '" + folder_name + "' at '" + folder_path.string()
-                           + "' deleted.\n";
+            std::cout << "Folder " + folder_name + " at " + folder_path.string()
+                           + " deleted.\n";
         }
         catch (const fs::filesystem_error& e)
         {
